requirements.txt
protoc-gen-openapiv2
langchain-pinecone
pinecone-notebooks
pinecone
langchain-community
langchain-google-genai
Flask
firebase-admin
PyPDF2
pinecone-client[grpc]
google-generativeai
python-dotenv
google-auth
google-cloud
